@@ -11,64 +11,69 @@
   "author": "Francesco Sullo <francesco@sullo.co>",
   "license": "MIT",
   "dependencies": {
-    "@ethersproject/contracts": "^5.1.0",
-    "@ethersproject/providers": "^5.1.0",
-    "body-parser": "^1.19.0",
+    "@ethersproject/contracts": "^5.8.0",
+    "@ethersproject/providers": "^5.8.0",
+    "@tanstack/react-query": "^5.83.0",
+    "@web3modal/wagmi": "^5.1.11",
+    "body-parser": "^1.20.3",
     "chalk": "^2.4.2",
-    "cookie-parser": "^1.4.5",
+    "cookie-parser": "^1.4.7",
     "cors": "^2.8.5",
     "discord.js": "^12.5.3",
     "dotenv": "^9.0.2",
     "eth-sig-util": "^3.0.1",
-    "ethereumjs-util": "^7.0.10",
-    "ethers": "^5.1.4",
-    "express": "^4.17.1",
+    "ethereumjs-util": "^7.1.5",
+    "ethers": "^5.8.0",
+    "express": "^4.21.2",
     "fs-extra": "^8.1.0",
     "ipfs-http-client": "^49.0.4",
-    "knex": "^0.21.12",
+    "knex": "^0.21.21",
     "local-storage": "^1.4.2",
     "lodash": "^4.17.21",
-    "query-string": "^7.0.1",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
-    "react-markdown": "^6.0.1",
-    "react-markdown-it": "^1.0.2",
-    "react-password-strength-bar": "^0.3.2",
-    "react-player": "^2.9.0",
-    "react-scroll": "^1.8.4",
-    "react-webcam": "^5.2.3",
+    "prop-types": "^15.8.1",
+    "query-string": "^7.1.3",
+    "react": "^18.3.1",
+    "react-bootstrap": "^1.6.8",
+    "react-dom": "^18.3.1",
+    "react-markdown": "^6.0.3",
+    "react-password-strength-bar": "^0.3.5",
+    "react-player": "^2.16.1",
+    "react-router-dom": "^7.7.0",
+    "react-scroll": "^1.9.3",
+    "react-webcam": "^5.2.4",
     "require-or-mock": "^0.2.1",
-    "simple-json-db": "^1.2.3",
+    "simple-json-db": "^1.3.0",
     "superagent": "^4.1.0",
     "video-react": "^0.14.1",
-    "web3modal": "^1.9.3"
+    "viem": "^2.32.1",
+    "wagmi": "^2.15.7",
+    "web3modal": "^1.9.12"
   },
   "devDependencies": {
-    "babel-core": "^6.26.3",
-    "babel-eslint": "^7.2.3",
-    "babel-loader": "^7.1.5",
-    "babel-plugin-react-transform": "^2.0.2",
-    "babel-polyfill": "^6.26.0",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-0": "^6.24.1",
-    "babel-register": "^6.26.0",
-    "css-loader": "^3.6.0",
-    "eslint": "^6.8.0",
-    "eslint-plugin-react": "^6.10.3",
-    "mocha": "^8.2.1",
-    "webpack": "^2.7.0"
+    "@vitejs/plugin-react": "^4.7.0",
+    "assert": "^2.1.0",
+    "buffer": "^6.0.3",
+    "crypto-browserify": "^3.12.1",
+    "eslint": "^8.57.0",
+    "eslint-plugin-react": "^7.34.0",
+    "mocha": "^10.3.0",
+    "process": "^0.11.10",
+    "stream-browserify": "^3.0.0",
+    "util": "^0.12.5",
+    "vite": "^7.0.5"
   },
   "scripts": {
     "start": "node .",
     "start:dev": "NODE_ENV=development node .",
-    "build": "webpack",
+    "build": "vite build",
+    "build:dev": "vite build --mode development",
+    "dev": "vite",
+    "preview": "vite preview",
     "docker-build": "docker buildx build --platform linux/amd64 -t everdragons2-app -f docker/Dockerfile .",
     "docker-run": "docker run --rm -p 6660:6660 everdragons2-app",
     "lint": "npm run lint:client && npm run lint:server",
     "lint:client": "eslint client/**/*.js",
-    "lint:server": "eslint server/**/*.js",
-    "dev": "npm run build && npm run start:dev"
+    "lint:server": "eslint server/**/*.js"
   },
   "eslintConfig": {
     "extends": "react-app"
