@@ -2,7 +2,7 @@ const express = require('express')
 const router = express.Router()
 const sigUtil = require('eth-sig-util')
 const db = require('../lib/Db')
-const Address = require('../../client/utils/Address')
+const Address = require('../utils/Address')
 const path = require('path')
 const fs = require('fs-extra')
 const {getContract} = require('../lib/utils')
